Initial set of unit tests.
There's pretty good coverage of valid PHP serialisations, which should be returned unmodified.

There are also quite a few tests for invalid serialisations, indicating how they are expected to be fixed.  However, this is a lot more patchy.  In particular, there is not really any test coverage of corrupted array serialisations yet.

More work to be done, but it's a good start.