Fixed test warning on PHP 8
Fixed a warning in the unit tests that only occurs on PHP 8:

  1) Warning
  The data provider specified for PHPSerialisationFixerTest::testValidSerialisations is invalid.
  Exception: Serialization of 'GdImage' is not allowed
  ~/tests/PHPSerialisationFixerTest.php:69

This is because the imagecreate() function no longer returns a reference, but a non-serialisable object.  I have fixed this by switching to using opendir() to create our resource, as this still returns a resource on all PHP versions.  I expect, in due course, this will end up returning an object, too, but until then this test is working properly again.

I also added a note to explain that the choice of function is arbitrary, to help me if I need to update it again.